move p-chain mock to external mocks dir
