+++ peers/avago_mocks/mock_platformvm_client.go
@@ -3,11 +3,11 @@
 //
 // Generated by this command:
 //
-//	mockgen -source=../avalanchego/vms/platformvm/client.go -destination=signature-aggregator/aggregator/mocks/mock_platformvm_client.go -package mocks
+//	mockgen -source=../avalanchego/vms/platformvm/client.go -destination=peers/avago_mocks/mock_platformvm_client.go -package avago_mocks
 //
 
-// Package mocks is a generated GoMock package.
-package mocks
+// Package avago_mocks is a generated GoMock package.
+package avago_mocks
 
 import (
 	context "context"
